Linking scriptchart with control accordion
