+++ js/components/DashTabs.js
@@ -5,7 +5,7 @@ import { CSSTransitionGroup } from "react-transition-group";
 import ScriptChart from "./ScriptChart";
 import MiradorContainer from "./MiradorContainer";
 
-import ChartAccordion from "./ControlAccordion";
+import ChartAccordion from "./ChartAccordion";
 
 import Tabs from "./Tabs";
 
@@ -13,26 +13,50 @@ import "./DashTabs.css";
 
 const ActiveTabContent = props => <div>{props.content}</div>;
 
-const tabList = [
-  {
-    name: "Scriptchart",
-    icon: "fa fa-table",
-    content: <ScriptChart />
-  },
-  {
-    name: "Manuscript Viewer",
-    icon: "fa fa-image",
-    content: <MiradorContainer />
-  }
-];
-
 class DashTabs extends React.Component {
   constructor(props) {
     super(props);
 
+    this.scriptchart = <ScriptChart toggleVector={this.toggleVector} />;
+    this.miradorContainer = <MiradorContainer />;
+    this.chartAccordion = <ChartAccordion toggleVector={this.toggleVector} />;
+
+    this.tabList = [
+      {
+        name: "Scriptchart",
+        icon: "fa fa-table",
+        content: this.scriptchart
+      },
+      {
+        name: "Manuscript Viewer",
+        icon: "fa fa-image",
+        content: this.miradorContainer
+      }
+    ];
+
     this.state = {
       activeTab: "Scriptchart"
     };
+  }
+
+  toggleVector( showOrHide, rowOrColumn, id, name ) {
+    console.log("Just got a request to " + showOrHide + " " + rowOrColumn + " " + id + " " + name);
+
+    if (showOrHide == "show") {
+      // Show (aka "unhide") requests only come from the accordion and are sent to the scriptchart
+      if (rowOrColumn == "column") {
+        //this.ref.scriptchart.onToggleColumn(id);
+      } else if (rowOrColumn == "row") {
+        //this.ref.scriptchart.onToggleRow(id);
+      }
+    } else if (showOrHide == "hide") {
+      // Hide requests only come from the scriptchart and are sent to the accordion
+      if (rowOrColumn == "column") {
+        //this.ref.accordion.onToggleManuscript( id, name);
+      } else if (rowOrColumn == "row") {
+        //this.ref.accordion.onToggleLetter( id, name);
+      }
+    }
   }
 
   changeActiveTab(tab) {
@@ -40,11 +64,11 @@ class DashTabs extends React.Component {
   }
 
   activeTabContent() {
-    const activeIndex = tabList.findIndex(tab => {
+    const activeIndex = this.tabList.findIndex(tab => {
       return tab.name === this.state.activeTab;
     });
 
-    return tabList[activeIndex].content;
+    return this.tabList[activeIndex].content;
   }
 
   render() {
@@ -54,14 +78,12 @@ class DashTabs extends React.Component {
           <div className="columns">
             <div className={"column"}>
               <Tabs
-                tabList={tabList}
+                tabList={this.tabList}
                 activeTab={this.state.activeTab}
                 changeActiveTab={this.changeActiveTab.bind(this)}
               />
             </div>
-            <div className={"column"}>
-              <ChartAccordion />
-            </div>
+            <div className={"column"}>{this.chartAccordion}</div>
           </div>
           <CSSTransitionGroup
             className="tabs-content"
